+++ src/answers/static.js
